Move the TUI lib to its own directory
